requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pytest-flask==1.3.0
plotly==5.21.0
pydub==0.25.1
requests==2.31.0
boto3==1.34.93